Создает модули
